+++ src/main/java/janggi/JanggiConsole.java
@@ -3,13 +3,13 @@ package janggi;
 import janggi.dao.piece.PieceHistoryManager;
 import janggi.dao.turn.TurnManager;
 import janggi.dto.PieceMove;
-import janggi.piece.PiecesFactory;
-import janggi.piece.board.Board;
-import janggi.piece.board.BoardOrder;
-import janggi.piece.players.Players;
-import janggi.piece.players.Team;
-import janggi.piece.players.Turn;
-import janggi.position.Position;
+import janggi.domain.piece.PiecesFactory;
+import janggi.domain.board.Board;
+import janggi.domain.board.BoardOrder;
+import janggi.domain.players.Players;
+import janggi.domain.players.Team;
+import janggi.domain.players.Turn;
+import janggi.domain.piece.position.Position;
 import janggi.utils.ExceptionHandler;
 import janggi.utils.StringParser;
 import janggi.view.InputView;
