+++ src/main/java/janggi/dao/piece/PieceDaoImpl.java
@@ -1,10 +1,10 @@
 package janggi.dao.piece;
 
+import janggi.domain.piece.Piece;
 import janggi.dto.PieceDto;
 import janggi.dto.PieceMove;
-import janggi.move.Piece;
-import janggi.piece.players.Team;
-import janggi.position.Position;
+import janggi.domain.players.Team;
+import janggi.domain.piece.position.Position;
 import janggi.utils.DBUtil;
 import java.sql.Connection;
 import java.sql.SQLException;
