+++ src/main/java/janggi/dao/piece/PieceHistoryManager.java
@@ -1,12 +1,12 @@
 package janggi.dao.piece;
 
-import janggi.move.Piece;
+import janggi.domain.piece.Piece;
 import janggi.dto.PieceDto;
 import janggi.dto.PieceMove;
-import janggi.piece.board.Board;
-import janggi.piece.players.Players;
-import janggi.piece.players.Team;
-import janggi.position.Position;
+import janggi.domain.board.Board;
+import janggi.domain.players.Players;
+import janggi.domain.players.Team;
+import janggi.domain.piece.position.Position;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
