+++ src/main/java/janggi/domain/piece/position/Position.java
@@ -1,6 +1,6 @@
-package janggi.position;
+package janggi.domain.piece.position;
 
-import janggi.move.Direction;
+import janggi.domain.piece.direction.Direction;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
