+++ src/main/java/janggi/dto/PieceDto.java
@@ -1,7 +1,7 @@
 package janggi.dto;
 
-import janggi.move.Piece;
-import janggi.piece.players.Team;
+import janggi.domain.piece.Piece;
+import janggi.domain.players.Team;
 
 public record PieceDto(Team team, Piece piece, int y, int x) {
 }
