+++ src/main/java/janggi/dto/PieceMove.java
@@ -1,8 +1,8 @@
 package janggi.dto;
 
-import janggi.move.Piece;
-import janggi.piece.players.Team;
-import janggi.position.Position;
+import janggi.domain.piece.Piece;
+import janggi.domain.players.Team;
+import janggi.domain.piece.position.Position;
 
 public record PieceMove(boolean wantRun, Team team, Piece piece, Piece caughtPiece,
                         Position currentPosition, Position arrivalPosition, boolean isCaught) {
