+++ src/test/java/janggi/dao/piece/FakePieceDao.java
@@ -2,8 +2,8 @@ package janggi.dao.piece;
 
 import janggi.dto.PieceDto;
 import janggi.dto.PieceMove;
-import janggi.piece.players.Team;
-import janggi.position.Position;
+import janggi.domain.players.Team;
+import janggi.domain.piece.position.Position;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
