+++ src/test/java/janggi/domain/piece/direction/MovementsTest.java
@@ -1,12 +1,8 @@
-package janggi.piece.direction;
+package janggi.domain.piece.direction;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import janggi.move.Direction;
-import janggi.move.Movement;
-import janggi.move.Movements;
-import janggi.move.Piece;
-import janggi.position.Position;
+import janggi.domain.piece.position.Position;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 
@@ -18,10 +14,11 @@ public class MovementsTest {
         final Position currentPosition = new Position(3, 3);
         final Position arrivalPosition = new Position(4, 3);
         final Movements movements = new Movements(List.of(new Movement(Direction.RIGHT)));
+        final boolean canMoveIterable = false;
 
         // When & Then
         assertThatThrownBy(
-                () -> movements.findMovements(currentPosition, arrivalPosition, Piece.HAN_SOLDIER))
+                () -> movements.findMovements(currentPosition, arrivalPosition, canMoveIterable))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("[ERROR] 적절한 움직임이 아닙니다.");
     }
