+++ src/test/java/janggi/domain/piece/direction/PathValidatorTest.java
@@ -1,12 +1,12 @@
-package janggi.piece.direction;
+package janggi.domain.piece.direction;
 
-import janggi.move.Direction;
-import janggi.move.Movement;
-import janggi.move.ObstacleStrategy;
-import janggi.move.PathValidator;
-import janggi.move.Piece;
-import janggi.piece.board.Board;
-import janggi.position.Position;
+import janggi.domain.piece.ObstacleTraversalRule;
+import janggi.domain.piece.direction.Direction;
+import janggi.domain.piece.direction.Movement;
+import janggi.domain.piece.position.PathValidator;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.position.Position;
+import janggi.domain.board.Board;
 import java.util.Map;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -23,11 +23,11 @@ class PathValidatorTest {
         // Given
         final Position currentPosition = new Position(3, 5);
         final Position arrivalPosition = new Position(4, 5);
-        final PathValidator pathValidator = new PathValidator(Piece.KING, ObstacleStrategy.BLOCK);
+        final PathValidator pathValidator = new PathValidator(ObstacleTraversalRule.BLOCK);
 
         // When & Then
         Assertions.assertThatThrownBy(
-                        () -> pathValidator.validatePath(currentPosition, arrivalPosition,
+                        () -> pathValidator.validatePath(currentPosition, arrivalPosition, true,
                                 new Board(Map.of(currentPosition, piece)), new Movement(Direction.DOWN)))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("[ERROR] 궁성 밖을 나갔습니다.");
