+++ src/test/java/janggi/domain/players/PlayersTest.java
@@ -1,13 +1,11 @@
-package janggi.piece;
+package janggi.domain.players;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import janggi.move.Piece;
-import janggi.piece.board.Board;
-import janggi.piece.players.Players;
-import janggi.piece.players.Team;
-import janggi.position.Position;
+import janggi.domain.piece.Piece;
+import janggi.domain.board.Board;
+import janggi.domain.piece.position.Position;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
