+++ src/test/java/janggi/domain/players/TurnTest.java
@@ -1,9 +1,7 @@
-package janggi.turn;
+package janggi.domain.players;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.piece.players.Team;
-import janggi.piece.players.Turn;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
